requirements/base.txt
django==4.2.1
psycopg2-binary==2.9.6
python-environ==0.4.54
django-cors-headers==4.0.0
django-jazzmin==2.6.0
# drf-yasg==1.20.5
djangorestframework==3.14.0
django-filter==23.2
django-rosetta==0.9.9
python-telegram-bot==13.15  # for telegram bot integration
django-image-uploader-widget
django-ckeditor
celery
redis
django-modeltranslation
django-recaptcha
django-auditlog
django-rosetta
django-phonenumbers
django-phonenumber-field
Pillow